refactor: merge PermissionHelper into PermissionManager and clean up WelcomeScreen
